+++ source/Public/Test-M365DSCPowershellDataFile.ps1
@@ -1,4 +1,5 @@
-Function Test-M365DSCPowershellDataFile {
+function Test-M365DSCPowershellDataFile
+{
     <#
     .Synopsis
     Tests the specified object against the information defined in the ExampleData
@@ -11,8 +12,8 @@ Function Test-M365DSCPowershellDataFile {
     data.
 
     .Parameter Test
-    Specifies the type of tests to run. Allowed values are:
-    'TypeValue', 'Required', 'Mandatory', 'TypeValue/Required', 'TypeValue/Mandatory', 'Required/Mandatory', 'TypeValue/Required/Mandatory'.
+    Specifies the Tests. Allowed values are:
+    'TypeValue','Required','Mandatory'
 
     .Parameter InputObject
     The object that contains the data object that needs to be tested.
@@ -20,355 +21,519 @@ Function Test-M365DSCPowershellDataFile {
     .Parameter MandatoryObject
     The object that contains the Mandatory data.
 
-    .Parameter Exclude_AvailableAsResource
-    All items that are available as a resource and have to be ignored.
+    .Parameter MandatoryAction
+    Action Type for test Mandatory 'Present', 'Absent'
 
-    .Parameter Exclude_Required
-    Required items have to be ignored.
+    .Parameter excludeAvailableAsResource
+    All items that are available as a resource and have to be ignored.  ( wildcards can be uses )
 
-    .Parameter Verbosity
+    .Parameter excludeRequired
+    Required items have to be ignored. ( no wildcards )
+
+    .Parameter Ignore_AllRequired
+    If specified, all required items will be ignored.
+
+    .Parameter pesterVerbosity
     Specifies the verbosity level of the output. Allowed values are:
     None', 'Detailed', 'Diagnostic'. Default is 'Detailed'.
 
-    .Parameter StackTraceVerbosity
+    .Parameter pesterStackTraceVerbosity
     Specifies the verbosity level of the output. Allowed values are:
     'None', 'FirstLine', 'Filtered', 'Full'. Default is 'Firstline'.
 
-    .Parameter PesterScript
+    .Parameter pesterShowScript
     If specified, the generated Pester script will be opened in an editor.
 
-    .Example
-        $Result = Test-M365DSCPowershellDataFile -Test TypeValue/Required -InputObject $InputObject
+    .Parameter pesterOutputObject
+    If specified, the executed Pester script result will returned.
 
     .Example
-        $Result = Test-M365DSCPowershellDataFile -Test TypeValue/Required/Mandatory `
-        -InputObject $InputObject `
-        -MandatoryObject $MandatoryObject `
-        -Exclude_AvailableAsResource *CimInstance, *UniqueID, *IsSingleInstance `
-        -Exclude_Required *CimInstance, *UniqueID `
-        -Verbosity Detailed `
-        -StackTraceVerbosity None `
-        -PesterScript `
-        -Verbose
+    $InputObject = Import-PSDataFile -path '%Filename%.psd'
+
+    Test-M365DSCPowershellDataFile -Test TypeValue, Required `
+    -InputObject $InputObject `
+    -excludeAvailableAsResource *CimInstance, *UniqueID, *IsSingleInstance `
+    -excludeRequired CimInstance, UniqueID `
+    -pesterShowScript
+
+    .Example
+    $InputObject = Import-PSDataFile -path '%Filename_InputObject%.psd1'
+    $MandatoryObject = Import-PSDataFile -path '$Filename_MandatoryObject%.psd1'
+
+    Test-M365DSCPowershellDataFile -Test TypeValue, Required, Mandatory `
+    -InputObject $InputObject `
+    -MandatoryObject $MandatoryObject `
+    -MandatoryAction Present `
+    -excludeAvailableAsResource *CimInstance, *UniqueID, *IsSingleInstance `
+    -excludeRequired CimInstance, UniqueID `
+    -pesterShowScript
+
 
     .NOTES
-        This function requires Modules: M365DSC.CompositeResources, ObjectGraphTools
-
+    This function requires Modules: M365DSC.CompositeResources, ObjectGraphTools
     #>
-    [CmdletBinding()]
-    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidUsingInvokeExpression', '', Justification = 'Need Invoke Expression for performance, input is validated')]
-    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseDeclaredVarsMoreThanAssignments', '', Justification = 'Bug powershell, Vars are declared')]
 
-    param (
+    [CmdletBinding()]
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseDeclaredVarsMoreThanAssignments', '', Justification = 'Bug powershell, Vars are declared')]
+    [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSReviewUnusedParameter', '', Justification = 'Vars are used in sub functions')]
+
+    param
+    (
         [Parameter(Mandatory = $true)]
-        [ValidateSet('TypeValue', 'Required', 'Mandatory', 'TypeValue/Required', 'TypeValue/Mandatory', 'Required/Mandatory', 'TypeValue/Required/Mandatory' )]
-        [String]$Test,
+        [ValidateSet('TypeValue', 'Required', 'Mandatory')]
+        [String[]]
+        $Test,
+
+        [Parameter(Mandatory = $True)]
+        [System.Object]
+        $InputObject,
+
+        [Parameter(Mandatory = $False )]
+        [System.Object]
+        $MandatoryObject,
+
+        [Parameter(Mandatory = $False )]
+        [ValidateSet('Present', 'Absent')]
+        [string]
+        $MandatoryAction,
+
+        [Parameter(Mandatory = $False)]
+        [String[]]
+        $excludeAvailableAsResource,
+
+        [Parameter(Mandatory = $False)]
+        [String[]]
+        $excludeRequired,
+
+        [Parameter(Mandatory = $False)]
+        [Switch]
+        $Ignore_AllRequired,
 
         [Parameter(Mandatory = $False)]
         [ValidateSet('None', 'Detailed', 'Diagnostic')]
-        [String]$Verbosity = 'Detailed',
+        [String]
+        $pesterVerbosity = 'Detailed',
 
         [Parameter(Mandatory = $False)]
         [ValidateSet('None', 'FirstLine', 'Filtered', 'Full')]
-        [String]$StackTraceVerbosity = 'FirstLine',
+        [String]
+        $pesterStackTraceVerbosity = 'FirstLine',
 
         [Parameter(Mandatory = $False)]
-        [Switch]$PesterScript
+        [Switch]
+        $pesterShowScript,
+
+        [Parameter(Mandatory = $False)]
+        [Switch]
+        $pesterOutputObject
     )
 
-    DynamicParam {
-        $RuntimeParameterDictionary = New-Object System.Management.Automation.RuntimeDefinedParameterDictionary
-
-        $AttributeCollection1 = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
-        $ParameterAttribute = New-Object System.Management.Automation.ParameterAttribute
-        $ParameterAttribute.Mandatory = $true
-        $AttributeCollection1.Add($ParameterAttribute)
-
-        $AttributeCollection2 = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
-        $ParameterAttribute = New-Object System.Management.Automation.ParameterAttribute
-        $ParameterAttribute.Mandatory = $False
-        $AttributeCollection2.Add($ParameterAttribute)
-
-        if ($Null -ne $Test ) {
-            $ParamName = 'InputObject'
-            $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParamName, [System.Object], $AttributeCollection1)
-            $RuntimeParameterDictionary.Add($ParamName, $RuntimeParameter)
-
-            $ParamName = 'Exclude_AvailableAsResource'
-            $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParamName, [String[]], $AttributeCollection2)
-            $RuntimeParameterDictionary.Add($ParamName, $RuntimeParameter)
-        }
-
-        if ($Test -like '*Mandatory*') {
-            $ParamName = 'MandatoryObject'
-            $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParamName, [System.Object], $AttributeCollection1)
-            $RuntimeParameterDictionary.Add($ParamName, $RuntimeParameter)
-        }
-
-        if ($Test -like '*Required*') {
-            $ParamName = 'Exclude_Required'
-            $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParamName, [String[]], $AttributeCollection2)
-            $RuntimeParameterDictionary.Add($ParamName, $RuntimeParameter)
-        }
-
-        return $RuntimeParameterDictionary
-    }
-
-    begin {
-
-        $InputObject = $PSboundParameters['InputObject']
-        $MandatoryObject = $PSboundParameters['MandatoryObject']
-        $Exclude_AvailableAsResource = $PSboundParameters['Exclude_AvailableAsResource']
-        $Exclude_Required = $PSboundParameters['Exclude_Required']
-
+    begin
+    {
         # Test if the ObjectGraphTools module is loaded and the class is available
-        if (-not ([System.Management.Automation.PSTypeName]'PSNode').Type) {
+        if (-not ([System.Management.Automation.PSTypeName]'PSNode').Type)
+        {
             Import-Module ObjectGraphTools -Force
         }
 
-        Class M365DSC_Reference_Values {
-            [string]$Type
-            [string]$Required
-            [String]$Description
-            [String]$ValidateSet
+        # Function to display elapsed and total time
+        function ShowElapsed
+        {
+            param
+            (
+                [Switch]
+                $reset
+            )
 
-            M365DSC_Reference_Values([string]$InputString) {
-                [array]$Result = $InputString.split('|').foreach{ $_.trim() }
-                $this.type = $Result[0]
-                $this.Required = $Result[1]
-                $this.Description = $Result[2]
-                $This.ValidateSet = $Result[3].foreach{ "'" + ($Result[3] -Replace '\s*\/\s*', "', '") + "'" }
+            if ($reset)
+            {
+                $script:totalTime = Get-Date; $script:elapsedTime = Get-Date
             }
-        }
-        function ShowElapsed {
-            param( [Switch]$Reset )
-            if ($Reset) {
-                $script:TotalTime = get-date
-                $script:ElapsedTime = get-date
+
+            if (-not $script:totalTime)
+            {
+                $script:totalTime = Get-Date
             }
-            if (-not $script:TotalTime ) { $script:TotalTime = get-date }
-            if (-not $script:ElapsedTime ) { $script:ElapsedTime = get-date }
-            $Result = '^ Elapsed: {0} TotalTime: {1} seconds' -f "$(($(get-date) - $ElapsedTime).TotalSeconds) Seconds".PadRight(30, " ") , ($(get-date) - $TotalTime).TotalSeconds
-            $script:ElapsedTime = Get-date
-            return $Result
+
+            if (-not $script:elapsedTime)
+            {
+                $script:elapsedTime = Get-Date
+            }
+
+            $result = '^ Elapsed: {0} TotalTime: {1} seconds' -f "$(($(Get-Date) - $elapsedTime).TotalSeconds) Seconds".PadRight(30, ' ') , ($(Get-Date) - $totalTime).TotalSeconds
+            $script:elapsedTime = Get-Date
+
+            return $result
         }
 
-        Function Space {
-            param([string]$Depth)
-            if ($Depth -ge 3) { $Depth = 3 }
-            Return " " * ($Depth)
-        }
+        # Function to test if a string is a valid GUID
+        function Test-IsGuid
+        {
+            [CmdletBinding()]
+            [OutputType([bool])]
+            param
+            (
+                [Parameter(Mandatory = $true, ValueFromPipeline = $True)]
+                [string]
+                $stringGuid
+            )
 
-        function Test-IsGuid {
-            [CmdletBinding()][OutputType([bool])]
-            param([Parameter(Mandatory = $true, ValueFromPipeline = $True)][string]$StringGuid)
-            process {
-                try {
-                    $ObjectGuid = [System.Guid]::Empty
-                    return [System.Guid]::TryParse($StringGuid, [ref]$ObjectGuid)
+            process
+            {
+                try
+                {
+                    $objectGuid = [System.Guid]::Empty
+                    return [System.Guid]::TryParse($stringGuid, [ref]$objectGuid)
                 }
-                catch {
+                catch
+                {
                     Write-Error "An error occurred while checking the GUID format: $_"
                     return $false
                 }
             }
         }
 
-        function Pester_Type_Should_Command {
-            [CmdletBinding()][OutputType([System.String])]
-            param([Parameter(Mandatory = $true)][string]$Type)
-            try {
-                switch ($Type) {
+        # Function to generate Pester "should" commands based on type
+        function Pester_Type_Should_Command
+        {
+            [CmdletBinding()]
+            [OutputType([System.String])]
+            param
+            (
+                [Parameter(Mandatory = $true)]
+                [string]
+                $type
+            )
+
+            try
+            {
+                switch ($type) {
                     'SInt32' { return "should -match '^\d+$' -because 'Must be a positive Integer'" }
                     'SInt64' { return "should -match '^\d+$' -because 'Must be a positive Integer'" }
                     'UInt32' { return "should -BeOfType 'Int'" }
                     'UInt64' { return "should -BeOfType 'Int'" }
                     'Guid' { return "Test-IsGuid | should -Be 'True'" }
-                    default { return "should -BeOfType '$Type'" }
+                    default { return "should -BeOfType '$type'" }
                 }
             }
-            catch {
+            catch
+            {
                 Write-Error "An error occurred while generating Pester 'should' assertion: $_"
                 return $null
             }
         }
 
-        Function Recursive_Node {
-            param ([psnode]$InputObject)
-            if ($InputObject.count -eq 0 ) { return }
+        # Class M365DSC reference values
+        class M365DSC_Reference_Values
+        {
+            [string]$type
+            [string]$required
+            [String]$description
+            [String]$validateSet
 
+            M365DSC_Reference_Values([string]$inputString) {
+                [array]$result = $inputString.split('|').foreach{ $_.trim() }
+                $this.type = $result[0]
+                $this.required = $result[1]
+                $this.description = $result[2]
+                $this.validateSet = $result[3].foreach{ "'" + ($result[3] -Replace '\s*\/\s*', "', '") + "'" }
+            }
+        }
 
-            $Childs = $InputObject | Get-ChildNode
-            # Required
-            if ($Test -like '*Required*') {
-                $LeafParent = $True
-                $Childs.foreach{ if ( $_ -isnot [PSLeafNode]) { $LeafParent = $false } }
-                if ($LeafParent) {
-                    $RefChilds = Invoke-Expression  $('$Obj_M365DataExample.{0}' -f $($node.Path) -replace "\[\d+\]", '[0]' )
-                    if ($RefChilds) {
-                        $Required_Nodes = $RefChilds.GetEnumerator().Where{ $_.value -like '*| Required |*' }.name
-                        foreach ($Required_Node in $Required_Nodes) {
-                            $Exclude_R = $Exclude_Required | where-object { $('{0}.{1}' -f $Node.Path, $Required_Node) -like $_ } #Check if is Exclude
-                            "     {0}{3}`$InputObject.{1}.{2} | should -not -BeNullOrEmpty -Because 'Required setting'" -f $(space $node.depth), $Node.path, $Required_Node, $(if ($Exclude_R.Length) { "#" })
-                        }
-                    }
+        function Create_Pesternodes_Mandatory
+        {
+            param
+            (
+                [psnode]$MandatoryObject
+            )
+
+            $MandatoryLeafs = $MandatoryObject | get-childnode -Recurse -Leaf
+            $MandatoryLeafs.foreach{
+                if ($MandatoryAction -eq 'Absent')
+                {
+                    "`$inputObject.{0} | should -BeNullOrEmpty -because 'Denied Mandatory Setting'" -f $_.Path
+                }
+                else
+                {
+                    "`$inputObject.{0} | should -Be {1} -Because 'Mandatory Setting'" -f $_.Path, $_.Value
+                }
+            }
+        }
+        function Create_PesterNode
+        {
+            param
+            (
+                [psnode]$nodeObject,
+                [switch]$recursive
+            )
+
+            $refNodePath = '{0}' -f $($nodeObject.Path) -replace '\[\d+\]', '[0]'
+            $objRefNode = $ht[$refNodePath]
+
+            # Exclude nodes that match the patterns defined in $excludeAvailableAsResource
+            foreach ($Exclude in  $excludeAvailableAsResource)
+            {
+                if ($nodeObject.Path -like $Exclude)
+                {
+                    "#`$inputObject.{0}" -f $nodeObject.Path
+                    return
                 }
             }
 
-            Foreach ($Node in $Childs) {
-                [Bool]$IsHashTable = $False
-                [bool]$IsArray = $False
+            # No Composite Resource available
+            if ($null -eq $objRefNode)
+            {
+                "`$inputObject.{0} | should -BeNullOrEmpty -because 'Not available as Composite Resource'" -f $nodeObject.Path
+                return
+            }
 
-                $Exclude_A = $Exclude_AvailableAsResource | where-object { $Node.path -like $_ } #Check if is Exclude
-                $HashKey = if ($Exclude_A.Length) { "#" } else { "" }
-
-                if ($Node.count -eq 0 ) { return }
-                if ( $node.Depth -le 3 ) { '   {0}It "{1}" {2}{{' -f $(space $node.depth), $Node.path, $(if ($Exclude_A.Length) { "-skip" }   ) }
-
-
-                if ((-not $IsArray) -and (-not $IsHashTable) -and ($Test -like "*TypeValue*")) {
-
-                    [Bool]$IsHashTable = Invoke-Expression  $('$Obj_M365DataExample.{0} -is [HashTable]' -f $($node.Path) -replace "\[\d+\]", '[0]' )
-                    if ($IsHashTable) { "     {0}{1}`$InputObject.{2} -is [HashTable] | should -BeTrue  " -f $(space $node.depth), $HashKey, $node.path }
-                    else {
-                        [bool]$IsArray = Invoke-Expression  $('$Obj_M365DataExample.{0} -is [Object[]]' -f $($node.Path) -replace "\[\d+\]", '[0]' )
-                        if ($IsArray) { "     {0}{1}`$InputObject.{2} -is [Array] | should -BeTrue  " -f $(space $node.depth), $HashKey, $node.path }
+            if ($nodeObject -is [psCollectionNode])
+            {
+                # Check Folder Type
+                if ($Typevalue)
+                {
+                    [Bool]$isHashTable = $( $objRefNode.valueType.name -eq 'HashTable')
+                    if ($isHashTable)
+                    {
+                        "`$inputObject.{0} -is [HashTable] | should -BeTrue" -f $nodeObject.Path
                     }
+                    else
+                    {
+                        "`$inputObject.{0} -is [Array] | should -BeTrue " -f $nodeObject.Path
+                    }
+                }
 
-                    $RefResult = Invoke-Expression  $('$Obj_M365DataExample.{0}' -f $($node.Path) -replace "\[\d+\]", '[0]' )
-                    if ($RefResult -is [string]) {
-                        $Obj_RefResult = [M365DSC_Reference_Values]::new($RefResult)
-                        if ($Obj_RefResult.type) {
-                            # Type Validation
-                            "     {0}{1}`$InputObject.{2} | {3}" -f $(space $node.depth), $HashKey, $node.path , $(Pester_Type_Should_Command $Obj_RefResult.type)
-                            if ($Obj_RefResult.ValidateSet) {
-                                # ValidationSet Validation
-                                "     {0}{1}`$InputObject.{2} | should -beIn {3}" -f $(space $node.depth), $HashKey, $node.path, $Obj_RefResult.ValidateSet
+                # Check for required
+                if ($Required)
+                {
+                    $objRequiredNodes = $htRequired["$($refNodePath)"]
+                    if ($objRequiredNodes)
+                    {
+                        foreach ($objRequiredNode in $objRequiredNodes)
+                        {
+                            if ($objRequiredNode.name -notin $excludeRequired )
+                            {
+                                "`$inputObject.{0}.{1} | should -not -BeNullOrEmpty -Because 'Required setting'" -f $nodeObject.path, $objRequiredNode.name
+                            }
+                            else
+                            {
+                                "#`$inputObject.{0}.{1} | should -not -BeNullOrEmpty -Because 'Required setting'" -f $nodeObject.path, $objRequiredNode.name
                             }
                         }
                     }
-                    if ($null -eq $RefResult ) {
-                        "     {0}{1}`$InputObject.{2} | should -BeNullOrEmpty -because 'Not available as Composite Resource'" -f $(space $node.depth), $HashKey , $node.path
+                }
+
+                # Recursively process child nodes if exsists
+                $childs = $nodeObject | Get-ChildNode
+                foreach ($node in $childs)
+                {
+                    if ($recursive)
+                    {
+                        Create_PesterNode -nodeObject $node -recursive
+                    }
+                }
+
+            }
+            else
+            {
+                # LeafNode
+                if ($Typevalue)
+                {
+                    $objRefNodeValue = [M365DSC_Reference_Values]::new($objRefNode.Value)
+                    # Type Validation
+                    if ($objRefNodeValue.type)
+                    {
+                        "`$inputObject.{0} | {1}" -f $nodeObject.path , $(Pester_Type_Should_Command $objRefNodeValue.type)
                     }
 
+                    # ValidationSet Validation
+                    if ($objRefNodeValue.validateSet)
+                    {
+                        "`$inputObject.{0} | should -beIn {1}" -f $nodeObject.path, $objRefNodeValue.validateSet
+                    }
                 }
-
-                if ($node -isnot 'PSLeafNode' ) { Recursive_Node $Node } # Recursive call
-                if ( $Node.Depth -eq 3 ) { '   {0}}}' -f $(space $node.depth) }
-
             }
         }
-
-        Function MandatoryObject {
-            param ($Workload)
-            if ($Test -like "*Mandatory*") {
-                '      It "{0} test Mandatory settings" {{' -f $Workload.name
-                $node_MandatoryObject = $MandatoryObject.NonNodeData.$($Workload.name) | Get-ChildNode -Leaf -Recurse
-                $node_MandatoryObject.ForEach{
-                    "        `$InputObject.NonNodeData.{0}.{1} | should -be '{2}'" -f $($Workload.name), $_.path, $_.value
-                }
-                '      }'
-            }
-        }
-
     }
-    process {
-        $script:TotalTime = get-date
-        ShowElapsed  -Reset | out-null
-        '-- Function Test-M365DSCPowershellDataFile ------------------------------ ' | Write-log
-        'Tests selected: {0}' -f $Test | write-log
-        'Exclude AvailableAsResource: {0}' -f ($Exclude_AvailableAsResource -join "; " ) | write-log
-        'ExcludeRequired: {0}' -f ($Exclude_Required -join "; ") | write-log
+    process
+    {
+        $Typevalue = $False
+        $Required = $False
+        $Mandatory = $False
 
-        'Load InputObject as type [psnode]' | Write-log
-        $Node_InputObject = $InputObject | get-node
-        ShowElapsed | write-log -Verbose
-
-        if ($MandatoryObject) {
-            'Load MandatoryObject as type [psnode] ' | Write-log
-            $Node_MandatoryObject = $MandatoryObject | get-node
-            ShowElapsed | write-log -Verbose
+        if ($Test -contains 'TypeValue')
+        {
+            $Typevalue = $true
+        }
+        if ($Test -contains 'Required')
+        {
+            $Required = $True
+        }
+        if ($Test -contains 'Mandatory')
+        {
+            $Mandatory = $True
+            if (-not $MandatoryObject -or -not $MandatoryAction)
+            {
+                throw 'Mandatory test needs the parameters MandatoryObject and MandatoryAction'
+            }
         }
 
-        'Load Example data from module M365DSC.CompositeResources' | Write-log
-        $Obj_M365DataExample = @(
-            if (Get-Module M365DSC.CompositeResources) {
-                Import-PSDataFile -Path ((((Get-Module M365DSC.CompositeResources)).path | Split-Path) + '\M365ConfigurationDataExample.psd1')
+        ShowElapsed -Reset | Out-Null
+        'Test(s) selected: {0}' -f $( $Test -join ', ') | Write-Log
+        'Load Example data from module M365DSC.CompositeResources' | Write-Log
+
+        $Module = Get-Module M365DSC.CompositeResources
+        if ($null -eq $Module)
+        {
+            $Module = Get-Module -ListAvailable M365DSC.CompositeResources
+            if ($null -eq $Module)
+            {
+                throw "Cannot find module 'M365DSC.CompositeResources'"
             }
-            else {
-                Import-PSDataFile -Path (((Get-Module -ListAvailable M365DSC.CompositeResources).path | Split-Path) + '\M365ConfigurationDataExample.psd1')
-            }
-        )
-        ShowElapsed | write-log -Verbose
+        }
+        $pathM365DataExample = Join-Path -Path ($Module.path | Split-Path) -ChildPath 'M365ConfigurationDataExample.psd1'
+        $objM365DataExample = Import-PSDataFile -path $pathM365DataExample.ToString()
 
-        # Generate pesterscript
-        'Generate pesterscript' | Write-Log
-        $Pester_Config = @(
-            "#Requires -Modules Pester"
+        ShowElapsed | Write-Log -Debug
 
-            "Describe '--- Check M365-DSC-CompositeResources configuration ---' {"
-            foreach ($Workload in $( $Node_InputObject | Get-node("NonNodeData") | Get-ChildNode)) {
-
-                if ($Test -like "*TypeValue*" -or $Test -like "*Required*") {
-                    "  Context 'Workload {0}' {{" -f $Workload.Name
-                    "    Context 'Type/Validatieset/Required' {"
-                    Recursive_Node -InputObject $($Workload | get-node)
-                    "    }"
+        'Create Hashtables for reference data ' | Write-Log
+        [hashtable]$ht = @{}
+        [hashtable]$htRequired = @{}
+        $nodeObject = $inputObject | Get-ChildNode
+        foreach ($node in $nodeObject)
+        {
+            foreach ($node in $($node | Get-ChildNode))
+            {
+                # Create Hashtabel Exampledata
+                $objM365DataExample | Get-node("$($node.path)") | Get-ChildNode -Recurse -IncludeSelf | ForEach-Object {
+                    $ht["$($_.path)"] = $_
+                    # Create HashTable Required
+                    if ($Required)
+                    {
+                        if ($_ -is [PSLeafnode])
+                        {
+                            if ($_.value -match '\| Required \|')
+                            {
+                                $parentPath = $_.parentnode.path.ToString()
+                                if (-not $htRequired.Contains("$parentPath"))
+                                {
+                                    $htRequired["$parentPath"] = [System.Collections.Generic.List[psnode]]::new()
+                                }
+                                $htRequired["$parentPath"].add( $_ )
+                            }
+                        }
+                    }
                 }
+            }
+        }
 
-                if ($Test -like "*Mandatory*") {
-                    "    Context 'Mandatory' {"
-                    MandatoryObject -Workload $($Workload)
-                    "    }"
+        ShowElapsed | Write-Log -Debug
+
+        'Create pester rules' | Write-Log
+
+        $pesterConfig = @(
+            '#Requires -Modules Pester'
+            'Describe ''--- Check M365-DSC-CompositeResources configuration ---'' {'
+            if ($Typevalue -or $Required)
+            {
+                '  Context ''AllNodes'' {'
+
+                '  }'
+                '  Context ''NonNodeData'' {'
+                foreach ($workload in ( $inputObject | get-node 'NonNodeData' | get-childnode ))
+                {
+                    '    Context ''{0}'' {{' -f $workload.Path
+                    '      It ''{0}'' {{' -f $workload.Path
+                    Create_PesterNode -nodeObject $workload | ForEach-Object { '        {0}' -f $_ }
+                    '      }'
+                    if ($workload -is [psCollectionNode])
+                    {
+                        foreach ($workloadFolder in ($workload | get-childnode ))
+                        {
+                            '      It ''{0}'' {{' -f $workloadFolder.Path
+                            Create_PesterNode -nodeObject $workloadFolder -recursive | ForEach-Object { '        {0}' -f $_ }
+                            '      }'
+                        }
+                    }
+                    '    }'
+                    '  }'
                 }
-                if ($Test -like "*TypeValue*" -or $Test -like "*Required*") { "  }" }
             }
-            "}"
+
+            if ($Mandatory)
+            {
+                '  Context ''Mandatory'' {'
+                '      It ''Mandatory'' {'
+                Create_Pesternodes_Mandatory -MandatoryObject $($MandatoryObject | get-node) | ForEach-Object { '        {0}' -f $_ }
+                '    }'
+                '  }'
+            }
+
+            '}'
         )
-        ShowElapsed | Write-Log -verbose
 
-        try {
-            # Ensure that $Pester_Config is defined
-            if (-not $Pester_Config) { throw "The variable `\$Pester_Config` is not defined." }
+        ShowElapsed | Write-Log -Debug
 
-            # Create a temporary Pester script file
-            $PesterScriptPath = [System.IO.Path]::ChangeExtension((New-TemporaryFile).FullName, '.tests.ps1')
-            $Pester_Config | Out-File -FilePath $PesterScriptPath -Force -Confirm:$false -Encoding ascii
+        # Remove empty lines from the $pesterConfig variable
+        $pesterConfig = $pesterConfig | Where-Object { $_.Trim() -ne '' }
 
-            # Open the Pester script in VSCode
-            if ($PesterScript) { psedit $PesterScriptPath }
+        try
+        {
+            # Ensure that $pesterConfig is defined before continuing
+            if (-not $pesterConfig) { throw "The variable `\$pesterConfig` is not defined." }
 
-            # Set parameters for the Pester container
-            $PesterParams = [ordered]@{ Path = $PesterScriptPath }
-            $PesterContainer = New-PesterContainer @PesterParams
+            # Create a temporary Pester script file for test execution
+            $pesterScriptPath = [System.IO.Path]::ChangeExtension((New-TemporaryFile).FullName, '.tests.ps1')
+            $pesterConfig | Out-File -FilePath $pesterScriptPath -Force -Confirm:$false -Encoding ascii
 
-            # Configure Pester settings
-            $PesterConfiguration = [PesterConfiguration]@{
-                Run    = @{ Container = $PesterContainer; PassThru = $true }
-                Should = @{ ErrorAction = "Continue" }
-                Output = @{ Verbosity = $Verbosity; StackTraceVerbosity = $StackTraceVerbosity }
+            # Open the generated Pester script in Visual Studio Code
+            if ($pesterShowScript) { psedit $pesterScriptPath }
+
+            # Set parameters for the Pester container that manages the test execution
+            $pesterParams = [ordered]@{ Path = $pesterScriptPath }
+            $pesterContainer = New-PesterContainer @pesterParams
+
+            # Execute the Pester tests and store the results in the $pesterResult variable
+            $pesterConfiguration = [PesterConfiguration]@{
+                Run    = @{ Container = $pesterContainer; PassThru = $true }
+                Should = @{ ErrorAction = 'Continue' }
+                Output = @{ Verbosity = $pesterVerbosity ; StackTraceVerbosity = $pesterStackTraceVerbosity }
             }
+
 
             # Execute Pester tests and store results
             'Execute pesterscript' | Write-Log
-            $PesterResult = Invoke-Pester -Configuration $PesterConfiguration
-            return $PesterResult
+            $pesterResult = Invoke-Pester -Configuration $pesterConfiguration
+            if ($pesterOutputObject) {
+                return $pesterResult
+            }
+
         }
-        catch {
+        catch
+        {
             Write-Error "An error occurred: $_"
         }
-        finally {
-            # Clean up the temporary file
-            if (Test-Path -Path $PesterScriptPath) {
-                Remove-Item -Path $PesterScriptPath -Force -ErrorAction SilentlyContinue
+        finally
+        {
+            # Remove the temporary Pester script file after the tests have been executed
+            if (Test-Path -Path $pesterScriptPath)
+            {
+                Remove-Item -Path $pesterScriptPath -Force -ErrorAction SilentlyContinue
             }
 
             ShowElapsed | Write-Log -Verbose
-
-            if ( $pesterResult.FailedCount -gt 0) { $Splat = @{ Failure = $true } } else { $Splat = @{} }
-            'Pester[{0}]  Tests:{1}  Passed:{2}  Failed:{3} ' -f $PesterResult.version, $PesterResult.TotalCount, $PesterResult.PassedCount, $pesterResult.FailedCount | write-log @splat
-
-            '-- End 2------------------------------------------------------------------ ' | Write-log
+            # Log the test results, with failure handling if any tests have failed
+            if ($pesterResult.FailedCount -gt 0)
+            {
+                $splat = @{ Failure = $true }
+            }
+            else
+            {
+                $splat = @{}
+            }
+            'Pester[{0}]  Tests:{1}  Passed:{2}  Failed:{3} ' -f $pesterResult.version, $pesterResult.TotalCount, $pesterResult.PassedCount, $pesterResult.FailedCount | Write-Log @splat
         }
+
+        ShowElapsed | Write-Log -Debug
     }
 }